Move J'eevee script to Dire Maul
