+++ src/scripts/kalimdor/feralas/dire_maul/npc_j_eevee.cpp
